submission/supplementary/code/core/numeric.py
"""Numerical computation utilities."""

from typing import Optional, Tuple, List, Dict
import numpy as np
from sympy import Expr
from .types import NumericValue, Energy, RealValue
from .errors import ComputationError

def integrate_phase_space(
    f: Expr,
    limits: Tuple[float, float],
    *,
    precision: Optional[float] = None
) -> NumericValue:
    """
    Integrate over phase space with error estimation.
    
    Args:
        f: Integrand expression
        limits: Integration limits
        precision: Required precision
        
    Returns:
        NumericValue: Integral value with uncertainty
    """
    try:
        value = np.trapz(f, limits)
        # Estimate numerical uncertainty
        uncertainty = abs(value) * (precision or 1e-6)
        return NumericValue(value, uncertainty)
    except Exception as e:
        raise ComputationError(f"Phase space integration failed: {e}")

def solve_field_equations(
    field_config: Dict,
    energy: Energy,
    *,
    max_iter: int = 1000
) -> NumericValue:
    """
    Solve field equations numerically.
    
    Args:
        field_config: Field configuration
        energy: Energy scale
        max_iter: Maximum iterations
        
    Returns:
        NumericValue: Solution with uncertainty
    """
    # Implementation here...